Create parse_cnf method to get problem info
The parse_cnf method takes a list of lines from a .cnf file in DIMACS
format and parses them to get the number of variables and clauses, and
the clauses in the problem. The program so far takes the cnf file
via standard input and displays the  number of variables and number
of clauses followed by a series of lists representing each clause.